Moves schema stuff into its own python file for organization
